refactor(public-rsvp): extract shared helpers into _public_rsvp_shared
